feat(cron): extract purge logic and schedule in server.ts
- Move purge body from the route handler (15d7ea9) into src/lib/cron/purge-tutorials.ts exporting runTutorialPurge() and reusable RETENTION_DAYS
- Slim the HTTP route to validate CRON_SECRET against the Authorization header, then delegate; same { deleted, files_removed } response shape
- Add node-cron + @types/node-cron
- In server.ts, register a daily 03:00 schedule that calls runTutorialPurge() in-process; failures logged with [purge-tutorials cron] tag and never crash the server
- Skip registration when NODE_ENV=test for deterministic tests
- Use a relative import in server.ts since ts-node does not resolve the @/* alias without extra setup
- Document CRON_SECRET in .env.example
