+++ .env.example
@@ -12,3 +12,6 @@ STORAGE_BACKEND=local
 SUPABASE_URL=
 SUPABASE_SERVICE_ROLE_KEY=
 SUPABASE_STORAGE_BUCKET=tutorial-materials
+
+# Required only for the manual cron HTTP trigger (POST /api/cron/purge-tutorials). Embedded scheduler in server.ts runs in-process and does not need this.
+CRON_SECRET=
