+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@supabase/supabase-js": "^2.105.4",
     "@types/busboy": "^1.5.4",
+    "@types/node-cron": "^3.0.11",
     "axios": "^1.15.0",
     "bcrypt": "^6.0.0",
     "busboy": "^1.6.0",
@@ -19,6 +20,7 @@
     "lucide-react": "^1.8.0",
     "next": "16.2.4",
     "next-pwa": "^5.6.0",
+    "node-cron": "^4.2.1",
     "pg": "^8.20.0",
     "react": "19.2.4",
     "react-dom": "19.2.4",
